third_party/WebKit/LayoutTests/inspector-protocol/timeline/timeline-start-bufferEvents-expected.txt

Recording started
    Event recorded: TimerFire
Recording stopped
Events:
FunctionCall
    TimeStamp
    TimerInstall
TimerFire
    FunctionCall
        TimeStamp
